project/requirements.txt
fastapi
uvicorn
langchain
langchain-community
google-generativeai
chromadb
PyPDF2
